[refactor] : SectorResponse Dto naming change
