+++ Ticket-Api/src/main/java/com/jnu/ticketapi/api/coupon/model/request/WaitingQueueRequest.java
@@ -1,17 +1,17 @@
 package com.jnu.ticketapi.api.coupon.model.request;
 
 
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Getter
 @Setter
 @AllArgsConstructor
-public class WaitingQueueRequestDto {
+public class WaitingQueueRequest {
     private String userId;
     private String couponTitle;
 }
