--- Ticket-Api/src/main/java/com/jnu/ticketapi/api/sector/model/internal/SectorDto.java
@@ -1,7 +1,0 @@
-package com.jnu.ticketapi.api.sector.model.internal;
-
-
-import lombok.Builder;
-
-@Builder
-public record SectorDto(Long sectorId, String sectorNum, String sectorName) {}
