+++ Ticket-Api/src/main/java/com/jnu/ticketapi/api/sector/model/request/SectorReadRequest.java
@@ -1,0 +1,7 @@
+package com.jnu.ticketapi.api.sector.model.request;
+
+
+import lombok.Builder;
+
+@Builder
+public record SectorReadRequest(Long sectorId, String sectorNum, String sectorName) {}
