+++ Ticket-Api/src/main/java/com/jnu/ticketapi/api/sector/service/SectorRegisterUseCase.java
@@ -1,8 +1,8 @@
 package com.jnu.ticketapi.api.sector.service;
 
 
-import com.jnu.ticketapi.api.sector.model.request.SectorReadResponse;
 import com.jnu.ticketapi.api.sector.model.request.SectorRegisterRequest;
+import com.jnu.ticketapi.api.sector.model.response.SectorReadResponse;
 import com.jnu.ticketcommon.annotation.UseCase;
 import com.jnu.ticketdomain.domains.coupon.domain.Sector;
 import com.jnu.ticketdomain.domains.coupon.out.SectorLoadPort;
@@ -54,6 +54,6 @@ public class SectorRegisterUseCase {
 
     public List<SectorReadResponse> findAll() {
         List<Sector> all = sectorLoadPort.findAll();
-        return SectorReadResponse.toSectorReadResponse(all);
+        return SectorReadResponse.toSectorReadResponses(all);
     }
 }
