+++ Ticket-Api/src/main/java/com/jnu/ticketapi/application/helper/Converter.java
@@ -8,7 +8,7 @@ import com.jnu.ticketapi.api.registration.model.request.TemporarySaveRequest;
 import com.jnu.ticketapi.api.registration.model.response.FinalSaveResponse;
 import com.jnu.ticketapi.api.registration.model.response.GetRegistrationResponse;
 import com.jnu.ticketapi.api.registration.model.response.TemporarySaveResponse;
-import com.jnu.ticketapi.api.sector.model.internal.SectorDto;
+import com.jnu.ticketapi.api.sector.model.request.SectorReadRequest;
 import com.jnu.ticketcommon.annotation.Helper;
 import com.jnu.ticketcommon.message.ResponseMessage;
 import com.jnu.ticketdomain.domains.coupon.domain.Sector;
@@ -19,11 +19,11 @@ import java.util.stream.Collectors;
 
 @Helper
 public class Converter {
-    public List<SectorDto> toSectorDto(List<Sector> sectorList) {
+    public List<SectorReadRequest> toSectorDto(List<Sector> sectorList) {
         return sectorList.stream()
                 .map(
                         sector ->
-                                SectorDto.builder()
+                                SectorReadRequest.builder()
                                         .sectorId(sector.getId())
                                         .sectorNum(sector.getSectorNumber())
                                         .sectorName(sector.getName())
@@ -32,7 +32,9 @@ public class Converter {
     }
 
     public GetRegistrationResponse toGetRegistrationResponseDto(
-            String email, Registration registration, List<SectorDto> sectorDtoList) {
+            String email,
+            Registration registration,
+            List<SectorReadRequest> sectorReadRequestList) {
         return GetRegistrationResponse.builder()
                 .email(email)
                 .name(registration.getName())
@@ -41,7 +43,7 @@ public class Converter {
                 .carNum(registration.getCarNum())
                 .isLight(registration.isLight())
                 .phoneNum(registration.getPhoneNum())
-                .sectors(sectorDtoList)
+                .sectors(sectorReadRequestList)
                 .selectSectorId(registration.getSector().getId())
                 .build();
     }
